create/update types
