+++ src/features/api/types/ConfentfulError.ts
@@ -1,0 +1,9 @@
+interface ConfentfulError {
+  status: number
+  statusText: string
+  message: string
+  request: unknown
+  details: { errors: Record<string, string> }
+}
+
+export type { ConfentfulError }
