+++ src/features/api/types/index.ts
@@ -1,1 +1,2 @@
 export * from './ApiResponse'
+export * from './ConfentfulError'
