+++ src/features/users/types/UserRole.ts
@@ -1,0 +1,5 @@
+export enum UserRole {
+  Admin = 'admin',
+  Racer = 'racer',
+  None = 'none',
+}
